Add Palindrome String checker program.
Java Palindrome String Verifier

The code is a part of my Data Structures & Algorithms (DSA) experience. It gets a string in its input and finds out it a palindrome or not with just a 2 pointer approach- by comparing characters at the end.

Strings 1 2 -> Palindrome Check  
Two-pointer approach  
Input: user given single-string  
Output: A Boolean answer on whether it is a palindrome  
Testing status: Tested and working

Learning Note:
It was exercised on completion of each basic challenge with arrays such as reversing an array. Intensified concepts relating to management and manipulation of strings and control flow in java.